Add -noecho to xmlchange commands in build/setup scripts.
Echoing of xmlchange commands to CaseStatus is meant to record changes made by
the user, which they would have to reapply if creating a new case. Thus, we want
to exclude commands that are run automatically, such as these. However, note
that automated tests (in cimetest) DO echo xmlchange commands, so you can have a
record of changes made for each test.

Requires cime0_3_27 or later.


git-svn-id: https://svn-ccsm-models.cgd.ucar.edu/pop2/branches/newtesting@70107 fe37f545-8307-0410-aea5-b40df96820b5