Update dt_count for ssp extensions
Set OCN_TRANSIENT differently for SSP534EXT and SSP585EXT compsets, and then
use ocn_transient when determining dt_count.

Note that I put placeholders in for setting ndep_data_type as well, I assume
those files will be made at some point? As far as I can tell, the other
variables that depend on OCN_TRANSIENT and are set based on ssps are not active
in this compset:

* abio_atm_d14c_filename
* abio_atm_co2_opt
* abio_atm_d14c_opt
* ciso_atm_d13c_opt
* ciso_atm_d14c_opt
* ciso_atm_d13c_filename
* ciso_atm_d14c_filename