Fix a bug in the iron flux computation
Parentheses were wrong in a function in ecosys_forcing_mod.F90, two parameters
were renamed, and two more were introduced to make the parameterization more
flexible.

Note that this commit breaks backwards compatibility -- there is no way to run
compute this parameterization as it was done in CESM 2.0 and 2.1. Because of
that, I also removed the old namelist values that had previously been triggered
when OCN_BGC_CONFIG=cesm2.0 (or cesm2.1).