marbl0.42.0 changed surface_flux_output type
MARBL was updated to be able to pass the 3D total chlorophyll field in addition
to just the surface chlorophyll; support for that feature changed the name of
several underlying datatypes and variables (e.g. replacing "surface flux
outputs" with the more general "outputs for GCM" because total chlorophyll
comes from the interior tendency computation rather than the surface flux
computation).